extract layout hook
